Update main and inputMatrix
Checking for a sufficient number of items

+++ kiselev.sergey/P3/inputMatrix.h
@@ -3,6 +3,6 @@
 #include <fstream>
 namespace kiselev
 {
-  std::istream& inputMatrix(std::istream& in, int* array, size_t m, size_t n, int& count_read);
+  std::istream& inputMatrix(std::istream& in, int* array, size_t m, size_t n, size_t& count_read);
 }
 #endif
